+++ Bittrex/JsonData/MarketJson.cs
@@ -10,7 +10,7 @@ namespace Bittrex.JsonData
             public string BaseCurrency { get; set; }
             public string MarketCurrencyLong { get; set; }
             public string BaseCurrencyLong { get; set; }
-            public double MinTradeSize { get; set; }
+            public decimal MinTradeSize { get; set; }
             public string MarketName { get; set; }
             public bool IsActive { get; set; }
             public DateTime Created { get; set; }
