+++ Bittrex/JsonData/MarketSummaryJson.cs
@@ -7,17 +7,17 @@ namespace Bittrex.JsonData
     public class MarketSummaryJson
     {
         public string MarketName { get; set; }
-        public double High { get; set; }
-        public double Low { get; set; }
-        public double Volume { get; set; }
-        public double Last { get; set; }
-        public double BaseVolume { get; set; }
+        public decimal High { get; set; }
+        public decimal Low { get; set; }
+        public decimal Volume { get; set; }
+        public decimal Last { get; set; }
+        public decimal BaseVolume { get; set; }
         public DateTime TimeStamp { get; set; }
-        public double Bid { get; set; }
-        public double Ask { get; set; }
+        public decimal Bid { get; set; }
+        public decimal Ask { get; set; }
         public int OpenBuyOrders { get; set; }
         public int OpenSellOrders { get; set; }
-        public double PrevDay { get; set; }
+        public decimal PrevDay { get; set; }
         public DateTime Created { get; set; }
     }
 }
