+++ Bittrex/JsonData/OrderBook/Buy.cs
@@ -5,7 +5,7 @@ namespace Bittrex.JsonData
     [JsonObject(Title = "Buy")]
     public class BuyJson
     {
-        public double Quantity { get; set; }
-        public double Rate { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal Rate { get; set; }
     }
 }
