+++ DAL/MarketSummaryPoco.cs
@@ -27,27 +27,27 @@ namespace DAL
         /// <summary>
         /// Максимальное значение в базовой валюте
         /// </summary>
-        public double High { get; set; }
+        public decimal High { get; set; }
         
         /// <summary>
         /// Минимальное значение в базовой валюте
         /// </summary>
-        public double Low { get; set; }
+        public decimal Low { get; set; }
         
         /// <summary>
         /// Объём торгов по маркету в основной валюте
         /// </summary>
-        public double Volume { get; set; }
+        public decimal Volume { get; set; }
         
         /// <summary>
         /// Последнее значение по маркету
         /// </summary>
-        public double Last { get; set; }
+        public decimal Last { get; set; }
         
         /// <summary>
         /// Объём торгов по маркету в базовой валюте
         /// </summary>
-        public double BaseVolume { get; set; }
+        public decimal BaseVolume { get; set; }
         
         /// <summary>
         /// ??? Временной штамп информации
@@ -57,12 +57,12 @@ namespace DAL
         /// <summary>
         /// ???
         /// </summary>
-        public double Bid { get; set; }
+        public decimal Bid { get; set; }
         
         /// <summary>
         /// ???
         /// </summary>
-        public double Ask { get; set; }
+        public decimal Ask { get; set; }
         
         /// <summary>
         /// ???
@@ -77,7 +77,7 @@ namespace DAL
         /// <summary>
         /// ???
         /// </summary>
-        public double PrevDay { get; set; }
+        public decimal PrevDay { get; set; }
         
         /// <summary>
         /// ???
